+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Latest
+### Added
+- Light/dark mode toggle.
+
+## [0.2.22] - 2023-01-25
 ### Changed
 - Upgraded dependencies.
 - Internal structure re-organisation and linting.